+++ Assets/Scripts/GameEngine/Character/Move/Effect/ActiveModifier.cs
@@ -2,12 +2,12 @@ public class ActiveModifier
 {
     public ModifierType Type { get; }
     public StatType Stat { get; }
-    public int Value { get; }
+    public float Value { get; }
     public ModifierOccurance Occurance { get; }
     public bool HasOccured { get; private set; }
     public int RemainingDuration { get; private set; }
 
-    public ActiveModifier(ModifierType type, StatType stat, int value, int duration)
+    public ActiveModifier(ModifierType type, StatType stat, float value, int duration)
     {
         Type = type;
         Stat = stat;
