+++ Assets/Scripts/GameEngine/Character/Move/Effect/Value/FlatValue.cs
@@ -7,7 +7,7 @@ public class FlatValue : IValue
         BaseValue = baseValue;
     }
 
-    public int GetValue(EffectContext context)
+    public float GetValue(EffectContext context)
     {
         return BaseValue;
     }
